Abstract CoreData boilerplate out of tests; use NSInMemoryStoreType
