+++ Tests/ContentfulPersistenceTests/PreseedingTests.swift
@@ -17,50 +17,10 @@ import CoreLocation
 
 class PreseededDatabaseTests: XCTestCase {
 
-    #if os(iOS) || os(macOS)
-    let storeURL = FileManager.default.urls(for: .documentDirectory,
-                                            in: .userDomainMask).last?.appendingPathComponent("Test.sqlite")
-    #elseif os(tvOS)
-    let storeURL = FileManager.default.urls(for: .cachesDirectory,
-    in: .userDomainMask).last?.appendingPathComponent("Test.sqlite")
-    #endif
-
-    func deleteCoreDataStore() {
-        guard FileManager.default.fileExists(atPath: self.storeURL!.absoluteString) == true else { return }
-
-        try! FileManager.default.removeItem(at: self.storeURL!)
-        try! FileManager.default.removeItem(at: append("-shm", to: self.storeURL!))
-        try! FileManager.default.removeItem(at: append("-wal", to: self.storeURL!))
-    }
-
-    lazy var managedObjectContext: NSManagedObjectContext = {
-        let modelURL = Bundle(for: type(of: self)).url(forResource: "Test", withExtension: "momd")
-        let mom = NSManagedObjectModel(contentsOf: modelURL!)
-        expect(mom).toNot(beNil())
-
-        let psc = NSPersistentStoreCoordinator(managedObjectModel: mom!)
-
-        do {
-            var store = try psc.addPersistentStore(ofType: NSSQLiteStoreType, configurationName: nil, at: self.storeURL!, options: nil)
-            expect(store).toNot(beNil())
-        } catch {
-            XCTAssert(false, "Recreating the persistent store SQL files should not throw an error")
-        }
-
-        var managedObjectContext = NSManagedObjectContext(concurrencyType: .mainQueueConcurrencyType)
-        managedObjectContext.persistentStoreCoordinator = psc
-        return managedObjectContext
-    }()
-
-    func append(_ string: String, to fileURL: URL) -> URL {
-        let pathString = fileURL.path.appending(string)
-        return URL(fileURLWithPath: pathString)
-    }
-
     var client: Client!
 
     lazy var store: CoreDataStore = {
-        return CoreDataStore(context: self.managedObjectContext)
+        return CoreDataStore(context: TestHelpers.managedObjectContext(forMOMInTestBundleNamed: "Test"))
     }()
 
     var syncManager: SynchronizationManager!
@@ -74,8 +34,6 @@ class PreseededDatabaseTests: XCTestCase {
 
         self.client = Client(spaceId: "dqpnpm0n4e75", accessToken: "95c33f933385aa838825526c5753f3b5a7e59bb45cd6b5d78e15bfeafeef1b13", persistenceIntegration: synchronizationManager)
         self.syncManager = synchronizationManager
-
-        self.deleteCoreDataStore()
     }
 
     let postPredicate = NSPredicate(format: "id == '1asN98Ph3mUiCYIYiiqwko'")
@@ -126,55 +84,12 @@ class PreseededDatabaseTests: XCTestCase {
 
 class MultiLocalePreseedTests: XCTestCase {
 
-    #if os(iOS) || os(macOS)
-    let storeURL = FileManager.default.urls(for: .documentDirectory,
-                                            in: .userDomainMask).last?.appendingPathComponent("LocalizationTest.sqlite")
-    #elseif os(tvOS)
-    let storeURL = FileManager.default.urls(for: .cachesDirectory,
-    in: .userDomainMask).last?.appendingPathComponent("LocalizationTest.sqlite")
-    #endif
-
-
-    func append(_ string: String, to fileURL: URL) -> URL {
-        let pathString = fileURL.path.appending(string)
-        return URL(fileURLWithPath: pathString)
-    }
-
-    func deleteCoreDataStore() {
-        guard FileManager.default.fileExists(atPath: self.storeURL!.absoluteString) == true else { return }
-
-        try! FileManager.default.removeItem(at: self.storeURL!)
-        try! FileManager.default.removeItem(at: append("-shm", to: self.storeURL!))
-        try! FileManager.default.removeItem(at: append("-wal", to: self.storeURL!))
-    }
-
-    lazy var managedObjectContext: NSManagedObjectContext = {
-        let modelURL = Bundle(for: type(of: self)).url(forResource: "LocalizationTest", withExtension: "momd")
-        let mom = NSManagedObjectModel(contentsOf: modelURL!)
-        expect(mom).toNot(beNil())
-
-        let psc = NSPersistentStoreCoordinator(managedObjectModel: mom!)
-
-        do {
-            // Store in memory so there is no caching between test methods.
-            var store = try psc.addPersistentStore(ofType: NSSQLiteStoreType, configurationName: nil, at: self.storeURL!, options: nil)
-            expect(store).toNot(beNil())
-        } catch {
-            XCTAssert(false, "Recreating the persistent store SQL files should not throw an error")
-        }
-
-        var managedObjectContext = NSManagedObjectContext(concurrencyType: NSManagedObjectContextConcurrencyType.privateQueueConcurrencyType)
-        managedObjectContext.persistentStoreCoordinator = psc
-        return managedObjectContext
-    }()
-
-
     var syncManager: SynchronizationManager!
 
     var client: Client!
 
     lazy var store: CoreDataStore = {
-        return CoreDataStore(context: self.managedObjectContext)
+        return CoreDataStore(context: TestHelpers.managedObjectContext(forMOMInTestBundleNamed: "LocalizationTest"))
     }()
 
     override func setUp() {
@@ -188,8 +103,6 @@ class MultiLocalePreseedTests: XCTestCase {
                              accessToken: "14d305ad526d4487e21a99b5b9313a8877ce6fbf540f02b12189eea61550ef34",
                              persistenceIntegration: synchronizationManager)
         self.syncManager = synchronizationManager
-
-        self.deleteCoreDataStore()
     }
 
     func testPreseededDatabaseHasRecordsForAllLocales() {
